pyproject.toml
[project]
name = "quick-crop"
version = "0.1.0"
description = "A lightweight desktop tool for fast, interactive image cropping optimized for Instagram."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.0",
    "pyqt6>=6.10.2",
]
